+++ core/designsystem/src/main/java/com/moya/funch/component/Button.kt
@@ -1,4 +1,4 @@
-package com.moya.funch.component.button
+package com.moya.funch.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -18,6 +18,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
@@ -25,15 +26,24 @@ import androidx.compose.ui.unit.dp
 import com.moya.funch.modifier.clickableSingle
 import com.moya.funch.modifier.neonSign
 import com.moya.funch.theme.FunchTheme
+import com.moya.funch.theme.Gray400
+import com.moya.funch.theme.Gray800
 import com.moya.funch.theme.Gray900
 import com.moya.funch.theme.Lemon500
 import com.moya.funch.theme.Lemon900
 import com.moya.funch.theme.Yellow500
+import com.moya.funch.theme.funchTypography
 
-private val DefaultContentPadding = PaddingValues(vertical = 16.dp, horizontal = 24.dp)
+enum class FunchButtonType(val shape: Shape, val contentVerticalPadding: Dp, val textStyle: TextStyle) {
+    Full(RoundedCornerShape(16.dp), 21.dp, funchTypography.sbt1),
+    Large(RoundedCornerShape(16.dp), 21.dp, funchTypography.sbt1),
+    Medium(RoundedCornerShape(16.dp), 16.dp, funchTypography.sbt2),
+    Small(RoundedCornerShape(12.dp), 12.dp, funchTypography.b),
+    XSmall(RoundedCornerShape(12.dp), 8.dp, funchTypography.b);
+}
 
 @Composable
-fun FunchSharingButton(
+fun FunchMainButton(
     buttonType: FunchButtonType,
     onClick: () -> Unit,
     modifier: Modifier = Modifier,
@@ -42,7 +52,7 @@ fun FunchSharingButton(
     contentHorizontalPadding: Dp = 0.dp,
     icon: @Composable () -> Unit = {},
 ) {
-    SharingButton(
+    FunchMainButton(
         modifier = modifier,
         onClick = onClick,
         enabled = enabled,
@@ -55,7 +65,7 @@ fun FunchSharingButton(
 }
 
 @Composable
-fun SharingButton(
+fun FunchMainButton(
     modifier: Modifier = Modifier,
     onClick: () -> Unit,
     enabled: Boolean = true,
@@ -87,7 +97,57 @@ fun SharingButton(
     }
 }
 
-@Preview(name = "button Type", showBackground = true, widthDp = 360)
+@Composable
+fun FunchSubButton(
+    buttonType: FunchButtonType,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+    enabled: Boolean = true,
+    text: String,
+    contentHorizontalPadding: Dp = 0.dp,
+    icon: @Composable () -> Unit = {},
+) {
+    val color = if (enabled) {
+        FunchTheme.colors.white
+    } else {
+        Gray400
+    }
+    FunchSubButton(
+        modifier = modifier,
+        onClick = onClick,
+        enabled = enabled,
+        shape = buttonType.shape,
+        contentPadding = PaddingValues(contentHorizontalPadding, buttonType.contentVerticalPadding)
+    ) {
+        Text(text = text, color = color, style = buttonType.textStyle)
+        icon()
+    }
+}
+
+@Composable
+fun FunchSubButton(
+    modifier: Modifier = Modifier,
+    onClick: () -> Unit,
+    enabled: Boolean = true,
+    shape: Shape = RoundedCornerShape(size = 16.dp),
+    contentPadding: PaddingValues = PaddingValues(0.dp),
+    content: @Composable RowScope.() -> Unit,
+) {
+    Box(
+        modifier = modifier
+            .clip(shape = shape)
+            .background(Gray800)
+            .clickableSingle(enabled = enabled, onClick = onClick)
+            .padding(contentPadding),
+        contentAlignment = Alignment.Center,
+    ) {
+        Row {
+            content()
+        }
+    }
+}
+
+@Preview(name = "main button size", showBackground = true, widthDp = 360)
 @Composable
 private fun Preview1() {
     FunchTheme {
@@ -98,35 +158,35 @@ private fun Preview1() {
                 .padding(horizontal = 20.dp, vertical = 20.dp),
             horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            FunchSharingButton(
+            FunchMainButton(
                 modifier = Modifier.fillMaxWidth(),
                 buttonType = FunchButtonType.Full,
                 onClick = { /*TODO*/ },
                 text = "Button"
             )
             Spacer(modifier = Modifier.padding(16.dp))
-            FunchSharingButton(
+            FunchMainButton(
                 buttonType = FunchButtonType.Large,
                 onClick = { /*TODO*/ },
-                contentHorizontalPadding = 131.dp,
+                contentHorizontalPadding = 120.dp,
                 text = "Button"
             )
             Spacer(modifier = Modifier.padding(16.dp))
-            FunchSharingButton(
+            FunchMainButton(
                 buttonType = FunchButtonType.Medium,
                 onClick = { /*TODO*/ },
                 contentHorizontalPadding = 60.dp,
                 text = "Button"
             )
             Spacer(modifier = Modifier.padding(16.dp))
-            FunchSharingButton(
+            FunchMainButton(
                 buttonType = FunchButtonType.Small,
                 onClick = { /*TODO*/ },
                 contentHorizontalPadding = 16.dp,
                 text = "Button"
             )
             Spacer(modifier = Modifier.padding(16.dp))
-            FunchSharingButton(
+            FunchMainButton(
                 buttonType = FunchButtonType.XSmall,
                 onClick = { /*TODO*/ },
                 text = "Button",
@@ -136,9 +196,58 @@ private fun Preview1() {
     }
 }
 
-@Preview(name = "enabled, disabled 버튼", showBackground = true, widthDp = 360)
+@Preview(name = "subButton - size", showBackground = true, widthDp = 360)
 @Composable
 private fun Preview2() {
+    FunchTheme {
+        Column(
+            Modifier
+                .background(Gray900)
+                .fillMaxWidth()
+                .padding(horizontal = 20.dp, vertical = 20.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            FunchSubButton(
+                modifier = Modifier.fillMaxWidth(),
+                buttonType = FunchButtonType.Full,
+                onClick = { /*TODO*/ },
+                text = "Button"
+            )
+            Spacer(modifier = Modifier.padding(16.dp))
+            FunchSubButton(
+                buttonType = FunchButtonType.Large,
+                onClick = { /*TODO*/ },
+                contentHorizontalPadding = 120.dp,
+                text = "Button"
+            )
+            Spacer(modifier = Modifier.padding(16.dp))
+            FunchSubButton(
+                buttonType = FunchButtonType.Medium,
+                onClick = { /*TODO*/ },
+                contentHorizontalPadding = 60.dp,
+                text = "Button"
+            )
+            Spacer(modifier = Modifier.padding(16.dp))
+            FunchSubButton(
+                buttonType = FunchButtonType.Small,
+                onClick = { /*TODO*/ },
+                contentHorizontalPadding = 16.dp,
+                text = "Button"
+            )
+            Spacer(modifier = Modifier.padding(16.dp))
+            FunchSubButton(
+                buttonType = FunchButtonType.XSmall,
+                onClick = { /*TODO*/ },
+                text = "Button",
+                contentHorizontalPadding = 12.dp
+            )
+        }
+    }
+}
+
+@Preview(name = "main button - enabled, disabled", showBackground = true, widthDp = 360)
+@Composable
+private fun Preview3() {
     FunchTheme {
         Column(
             Modifier
@@ -146,7 +255,7 @@ private fun Preview2() {
                 .fillMaxWidth()
                 .padding(horizontal = 20.dp, vertical = 20.dp),
         ) {
-            SharingButton(contentPadding = PaddingValues(vertical = 16.dp, horizontal = 24.dp), onClick = { }) {
+            FunchMainButton(contentPadding = PaddingValues(vertical = 16.dp, horizontal = 24.dp), onClick = { }) {
                 Text(
                     text = "Button",
                     style = FunchTheme.typography.sbt1,
@@ -156,7 +265,7 @@ private fun Preview2() {
             }
             Spacer(modifier = Modifier.padding(16.dp))
 
-            SharingButton(
+            FunchMainButton(
                 enabled = false,
                 contentPadding = PaddingValues(vertical = 16.dp, horizontal = 24.dp),
                 onClick = { }) {
@@ -167,6 +276,35 @@ private fun Preview2() {
                     textAlign = TextAlign.Center,
                 )
             }
+        }
+    }
+}
+
+@Preview(name = "subButton - enabled, disabled ", showBackground = true, widthDp = 360)
+@Composable
+private fun Preview4() {
+    FunchTheme {
+        Column(
+            Modifier
+                .background(Gray900)
+                .fillMaxWidth()
+                .padding(horizontal = 20.dp, vertical = 20.dp),
+        ) {
+            FunchSubButton(
+                buttonType = FunchButtonType.Medium,
+                onClick = { /*TODO*/ },
+                contentHorizontalPadding = 60.dp,
+                text = "Button"
+            )
+
+            Spacer(modifier = Modifier.padding(16.dp))
+            FunchSubButton(
+                enabled = false,
+                buttonType = FunchButtonType.Medium,
+                onClick = { /*TODO*/ },
+                contentHorizontalPadding = 60.dp,
+                text = "Button"
+            )
         }
     }
 }
